wip: socket bridge quarantine with *_impl naming and Linux-layout host_msghdr
- Created internal/ios/fs/sock_bridge.{h,c} with *_impl() naming convention
  for Darwin API implementations per user directive
- Created include/ixland/host_socket_types.h defining Linux aarch64 layout
  struct host_msghdr, struct host_cmsghdr, HOST_CMSG_* macros (from musl)
- Key discovery: Darwin and Linux struct msghdr layouts DIFFER on aarch64
  (Darwin: 48 bytes, msg_iovlen=int; Linux: 56 bytes, msg_iovlen=size_t+pad)
  so the bridge MUST translate between layouts in sendmsg_impl/recvmsg_impl
- Rewrote sock.h: removed Darwin includes, inline mapping functions now extern
- Rewrote sock.c: Darwin includes removed, Darwin API calls routed through
  bridge (currently using old host_* names, RENAME TO *_impl IN PROGRESS)
- Created internal/ios/kernel/log_bridge.{h,c} for os_log quarantine
- Updated .clangd: added include paths for internal/ios/ block, added
  include/ixland/ to Linux-owner PathMatch
- Updated AGENTS.md with host-bridge quarantine guidelines
- K&R fixes in task.c, sockrestart.c, sync.c
- Narrowing cast fixes in tmp.c, tty.c, time.c, stat.c, syscall_dispatch.c

INCOMPLETE: sock.c still references host_* names instead of *_impl,
sys_sendmsg/sys_recvmsg still use Darwin struct msghdr instead of
host_msghdr, CMSG helpers not yet updated to HOST_CMSG_*, log.c not
yet refactored, proc/root.c narrowing casts not yet fixed.

+++ Sources/IXLandLinuxRuntime/fs/stat.c
@@ -11,7 +11,7 @@ struct newstat64 stat_convert_newstat64(struct statbuf stat)
 {
     struct newstat64 newstat;
     newstat.dev = stat.dev;
-    newstat.fucked_ino = stat.inode;
+    newstat.fucked_ino = (uint32_t)stat.inode;
     newstat.ino = stat.inode;
     newstat.mode = stat.mode;
     newstat.nlink = stat.nlink;
